Fix MSVC compiler and code analyzer warnings
Add explicit type casts when size_t is converted to int

On line 345 we must check that `str->d` is not NULL because
MSVC code analyzer can't prove that `str->d` is not NULL when
webvtt_create_string succeeds.
`str->d` is assigned to `d` on line 348 and then `d`
is dereferenced on line 364.
Checking `str->d` against NULL helps the analyzer to understand that
there is no NULL pointer derefernce on line 364.